Moved stream utils to separate package
Added missing license headers
