+++ pkg/messaging/v1/util.go
@@ -16,8 +16,6 @@ package v1
 import (
 	"context"
 	"encoding/base64"
-	"fmt"
-	"io"
 	"net/http"
 	"strconv"
 	"strings"
@@ -461,54 +459,4 @@ func WithCustomGRPCMetadata(ctx context.Context, md map[string]string) context.C
 	}
 
 	return ctx
-}
-
-/*!
-Adapted from the Go 1.18.3 source code
-Copyright 2009 The Go Authors. All rights reserved.
-License: BSD (https://github.com/golang/go/blob/go1.18.3/LICENSE)
-*/
-
-// TeeReadCloser is like io.TeeReader but returns a stream that can be closed, and which closes the readable stream.
-func TeeReadCloser(r io.ReadCloser, w io.Writer) io.ReadCloser {
-	return &teeReadCloser{
-		r: r,
-		w: w,
-	}
-}
-
-// teeReadCloser is an io.TeeReader that also implements the io.Closer interface to close the readable stream.
-type teeReadCloser struct {
-	r io.ReadCloser
-	w io.Writer
-}
-
-// Close implements io.Closer.
-func (t *teeReadCloser) Close() error {
-	var errR, errW error
-	errR = t.r.Close()
-	// w does not need to implement the io.Closer interface
-	if w, ok := t.w.(io.Closer); ok {
-		errW = w.Close()
-	}
-
-	if errR != nil && errW != nil {
-		return fmt.Errorf("failed to close r stream with error='%v' and failed to close w stream with error='%v'", errR, errW)
-	} else if errR != nil {
-		return fmt.Errorf("failed to close r stream with error='%v'", errR)
-	} else if errW != nil {
-		return fmt.Errorf("failed to close w stream with error='%v'", errW)
-	}
-	return nil
-}
-
-// Read from the R stream and tee it into the w stream.
-func (t *teeReadCloser) Read(p []byte) (n int, err error) {
-	n, err = t.r.Read(p)
-	if n > 0 {
-		if n, err := t.w.Write(p[:n]); err != nil {
-			return n, err
-		}
-	}
-	return n, err
 }
